pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kiui"
version = "0.3.1"
description = "A toolkit for 3D vision"
readme = { file = "README.md", content-type = "text/markdown" }
license = { file = "LICENSE" }
authors = [
  { name = "kiui", email = "ashawkey1999@gmail.com" },
]
urls = { Homepage = "https://github.com/ashawkey/kiuikit" }

requires-python = ">=3.8"

dependencies = [
  "lazy_loader",
  "varname",
  "objprint",
  "rich",
  "pyyaml",
]

[project.optional-dependencies]
full = [
  "torch", # keep this so the doc can be built successfully
  "tqdm",
  "numpy",
  "scipy",
  "scikit-image",
  "scikit-learn",
  "pandas",
  "trimesh",
  "pygltflib",
  "matplotlib",
  "opencv-python",
  "imageio",
  "imageio-ffmpeg",
  "dearpygui",
  "tyro",
  "viser",
  "pymeshlab",
  "huggingface_hub",
  "PyMCubes",
  "fbxloader",
  "envlight",
  "prompt_toolkit",
  "pygments",
  "openai",
]

[project.scripts]
kire = "kiui.render:main"
kisr = "kiui.sr:main"
kivi = "kiui.video:main"
kiss = "kiui.sys:main"
kia = "kiui.agent.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["kiui*"]
